requirements.txt
dds_cloudapi_sdk==0.2.1
gradio==4.36.1
h5py==3.8.0
httpx==0.27.0
imageio==2.33.1
imagesize==1.4.1
kornia==0.7.2
matplotlib==3.7.1
mmdet==3.3.0
mmengine==0.8.2
mmcv==2.0.1
ninja==1.11.1
nltk==3.8.1
open3d==0.16.0
openai==1.35.1
opencv_python==4.8.0.74
opencv_python_headless==4.9.0.80
pandas==2.0.2
plotly==5.15.0
plyfile==1.0.3
pydegensac==0.1.2
pytest==8.2.2
PyYAML==6.0.1
segment_anything==1.0
sphinx_rtd_theme==2.0.0
supervision==0.21.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
ultralytics
numpy
setuptools
